requirements.txt
Flask
gunicorn
line-bot-sdk
requests
psycopg2
numpy
Pillow
azure-ai-formrecognizer
sshtunnel
pymongo